+++ common/src/main/java/com/evandev/redomesticate/api/IPetbedDataEntity.java
@@ -1,8 +1,13 @@
 package com.evandev.redomesticate.api;
 
 import net.minecraft.nbt.CompoundTag;
+import net.minecraft.resources.ResourceLocation;
+
+import java.util.Map;
 
 public interface IPetbedDataEntity {
+    Map<ResourceLocation, Integer> redomesticate$getCachedEnchants();
+    void redomesticate$setCachedEnchants(Map<ResourceLocation, Integer> enchants);
 
     CompoundTag redomesticate$getEntityData();