+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.7.0] - 2026-05-31
+
+### Added
+
+- Added config options to enable/disable every enchantment from the mod.
+
+### Fixed
+
+- Performance improvements.
+- Fixed tamed Foxes not sitting when trinary commands are disabled.
+
 ## [1.6.1] - 2026-05-30
 
 ### Added